+++ hmjz/Chat/Class/Contacts/ContactsViewController.m
@@ -572,7 +572,7 @@
     }
     else
     {
-        NSString *tmpStr = [NSString stringWithFormat:@"%i", count];
+        NSString *tmpStr = [NSString stringWithFormat:@"%li", (long)count];
         CGSize size = [tmpStr sizeWithFont:self.unapplyCountLabel.font constrainedToSize:CGSizeMake(50, 20) lineBreakMode:NSLineBreakByWordWrapping];
         CGRect rect = self.unapplyCountLabel.frame;
         rect.size.width = size.width > 20 ? size.width : 20;
